requirements.txt
flask==2.2.2
flask-sqlalchemy==3.0.3
mysqlclient==2.1.1
Flask-Login==0.6.2
SQLAlchemy==2.0.1
python-dotenv==0.21.1